modularized the React code
