+++ app/js/lib/index.jsx
@@ -1,0 +1,7 @@
+const React = require('react');
+
+module.exports = {
+  Landing: require('./landing.jsx'),
+  Login: require('./login.jsx'),
+  Register: require('./register.jsx')
+};
