+++ app/js/lib/landing.jsx
@@ -1,0 +1,14 @@
+const React = require('react');
+
+const Landing = new React.createClass({
+  render: function() {
+    return (
+      <section>
+        <button>LOGIN</button>
+        <button>REGISTER</button>
+      </section>
+    );
+  }
+});
+
+module.exports = Landing;
